First stage of major repository restructure.
