+++ musemonitor/plot/shaders.py
@@ -1,0 +1,60 @@
+"""
+
+"""
+
+VERT_SHADER = """
+#version 120
+// y coordinate of the position.
+attribute float a_position;
+// row, col, and time index.
+attribute vec3 a_index;
+varying vec3 v_index;
+// 2D scaling factor (zooming).
+uniform vec2 u_scale;
+// Size of the table.
+uniform vec2 u_size;
+// Number of samples per signal.
+uniform float u_n;
+// Color.
+attribute vec3 a_color;
+varying vec4 v_color;
+// Varying variables used for clipping in the fragment shader.
+varying vec2 v_position;
+varying vec4 v_ab;
+void main() {
+    float nrows = u_size.x;
+    float ncols = u_size.y;
+    // Compute the x coordinate from the time index.
+    float x = -1 + 2*a_index.z / (u_n-1);
+    vec2 position = vec2(x - (1 - 1 / u_scale.x), a_position);
+    // Find the affine transformation for the subplots.
+    vec2 a = vec2(1./ncols, 1./nrows)*.9;
+    vec2 b = vec2(-1 + 2*(a_index.x+.5) / ncols,
+                  -1 + 2*(a_index.y+.5) / nrows);
+    // Apply the static subplot transformation + scaling.
+    gl_Position = vec4(a*u_scale*position+b, 0.0, 1.0);
+    v_color = vec4(a_color, 1.);
+    v_index = a_index;
+    // For clipping test in the fragment shader.
+    v_position = gl_Position.xy;
+    v_ab = vec4(a, b);
+}
+"""
+
+FRAG_SHADER = """
+#version 120
+varying vec4 v_color;
+varying vec3 v_index;
+varying vec2 v_position;
+varying vec4 v_ab;
+void main() {
+    gl_FragColor = v_color;
+    // Discard the fragments between the signals (emulate glMultiDrawArrays).
+    if ((fract(v_index.x) > 0.) || (fract(v_index.y) > 0.))
+        discard;
+    // Clipping test.
+    vec2 test = abs((v_position.xy-v_ab.zw)/v_ab.xy);
+    if ((test.x > 1))
+        discard;
+}
+"""
