+++ musemonitor/transform/defaults.py
@@ -1,0 +1,16 @@
+"""
+
+"""
+
+EEG_BANDS = {
+    "delta": (0, 4),
+    "theta": (4, 8),
+    "low_alpha": (8, 10),
+    #"med_alpha": (9, 11),
+    "hi_alpha": (10, 12),
+    "low_beta": (12, 21),
+    "hi_beta": (21, 30),
+    # "alpha": (8, 12),
+    # "beta": (12, 30),
+}
+"""dict: Default EEG bands, ex. for feature separation."""
